pyproject.toml
[tool.poetry]
name = "collegecostcalculator"
version = "0.1.0"
description = ""
authors = ["Sam Crabtree"]

[tool.poetry.dependencies]
python = "^3.10"
requests = "^2.28.0"
pandas = "^1.4.3"
python-dotenv = "^0.20.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
